Svelte Testing v2
This is a very updated version of my testing app. I added in a few new elements just to play around. I then edited these elements by importing a CSS file into +layout.svelte which allows me to change my elements as I pleased. I also played around with some JavaScript and created a counter with it and displayed in with HTML. I also introduced FlowBite into my svelte app which gave me access to a CSS library built specifically for svelte. I used this to add an alert and notification but I did not edit these. I also attempted to play around with the dark mode which currently only changes the FlowBite components and the colour of an h1 element.