Svelte Testing v3
Added in the css styles sheet for FlowBite allowing me to easily use it. Had to change a few file names to add the new css file. Deleted my previous styles in +layout.svelte and added them in as a class in the main +page.svelte file. Dark mode is also somewhat working as I can change elements by clicking the dark mode button which is essentially what it's supposed to do.