Svelte Testing v5
I have made a few major changes to my code as well as fixed all of my bugs generated while I made these changes. I have not updated in a while due to multiple issues I was facing at that time. I added in a streaming button that uses the devices camera through html and js. I then played around with html inputs and see if I could upload/take pictures to my app and it worked. My phone did not work because google chrome and android 14 cause a couple issues when it came to cameras. I spent a while solving this bug until I found a relatively recent forum page that said it was a bug. I also made a script that set a variable to an uploaded image's dataurl and so my testing for now if complete