Svelte Testing v6
Added in a function that that displays the input of a variable. I also changed the image input a little bit and played around with the variables. Created a button that calls the function display that I just created. This backup is in case my code and Owen's code does not join as expected.