App Design v1
Created a very simple design for my app which included most of its functionalities. The design consisted of the index page where the user can select images which will take them to the chats page where they will talk to ChatGPT about the image. Also added in a design for history and an preferences page all of which are connected with a three dots "more" button