App Design v2
This is my second version of my app. The main change I made was switching out the three dots "more" button to a navigation bar  on the footer. The user can either click the icon in the navigation bar in the footer or swipe left and right to get to their desired page. I also made the white background on light mode a bit duller to reduce the contrast between the background and the text. I also changed the position of certain elements slightly to be more centred and in-line with other elements