App Design v3
The main changes that took place between version 2 and 3 was the navigation bar in the footer and the index page. The index page was combined with the chats page to make a single index page where the user inserts an image and talks to ChatGPT. This reduced the number of pages from 4 to 3 and thus, a new design for the navigation bar was created. It now consists of 3 circular buttons inspired by Egg, Inc. Also added a microphone button for voice to text and a speaker icon for text to voice. These were supposed to be added as early as version 1 but I ended up forgetting