App Design v4
I changed the input and display images from squares into rectangles. This is because very few images will be squares and will instead be in a 3:4 ratio of width to height. I also changed the input image example to a more natural one that I took myself as the example image drew a lot of attention away from the design itself. I once again remodelled the buttons in the footer. I noticed that the main reason they looked off was because I was playing around with the border when I should've left it off. I switched some of the colours around as a result applied a very thin border to the light mode buttons to help it stand out against the background but not the rest of the app. I also switched up the sliders in the preferences page to be more consistent and applied another thin border on the blue sliders in light mode for the same reasons as the buttons in lights mode