App Design v5
I changed up the buttons again by making whatever page is active more obvious by increasing the contrast between the icon and its background. Also added a small horizontal line to divide the main page and the buttons as they had little relation in terms of information. Also added a trash icon next to the "clear all browsing data" as it was an opportunity to make a match between the system and real world. For the chats page, I added another box around all of the text that lightly contrasted off the background to show  which part of the page was the main part.