Ai Tourist v2 + v3
Merged our different repositories in a way which affected how I uploaded different versions of my app and thus v2 and v3 will be described simultaneously here. v2 added an input  file where users could upload an image and it would be displayed there. This image would also be converted into a data URL and stored as a variable for further use. Also added in a text input field which entered text is then displayed through HTML. This is preparation for my partners code. v3 added in three buttons for history, chats and settings respectively. Images have also been added in a new folder for these buttons. Inactive buttons will be done as .svelte components while active buttons will be hardcoded as they only occur once. Combined my app with my partners prototype and connected up the different variables and html. When an image is added through the image input and text is added through the text input, chatGPT will know respond (slowly) to the inputs and it will be displayed accordingly.