Ai Tourist v6
Did a little bit of extra research on API's and how they worked. As a result I managed to get Text to Speech working using OpenAI's tts-1 model. I used the documentation and was able to generate audio from text easily. I also wanted to run this on the server side as it hides the api key rather than exposing it like the ChatGPT one for responses. The difficult part was getting the specific text that the user wanted to turn to speech and sending it through the function. I came across event delegation and used that as a means to identify exactly which speaker icon was clicked. I also added extra data attributes as a means to select the text associated with the speaker icon. This allowed me to create text to speech