Ai Tourist v7
Was originally going to use live audio transcription but it proved to be very complicated. As a result I instead recorded audio and then stored that audio as an mp3 file. The mp3 file is then sent to OpenAi's whisper, transcribed into text, and appended to the contents of the text input. Also fixed the image upload as user will upload many different photo sizes. The  div that the image over the file upload changes size based on the dimensions of the image provided.