Ai Tourist v8
tidied up the code and commenting across both the main +page.svelte file and other server files. Began creating a server file for the actual conversation with ChatGPT, This is because it is a resource intensive task which should be done server side to free up devices resources. Also allows me to hide the api security key for better security. Is not working currently and has been parked for the future and I have started some error checking for inputs. I am currently adding little popup notifications when errors are made