Ai Tourist v10
added some error checking things to text to voice. Originally, multiple instances of recording would overlap over each other if multiple buttons were clicked. Now, if a transcription is being processed, it will be aborted and if audio is playing, it will be paused. This ensures that the last button clicked is the one transcribed and is the only audio playing