Ai Tourist v12
Moved most of my partners code that generated the AI responses to a server file for improved security. Fixed up all errors picked up by ESLint. Fully/mostly comment code; there may still be parts I missed as this is not the final version. Also added a similar animation for opacity and scale to the speaker icons so they now visually show that they were clicked on. When variables are incremented, they are now added to an array and looped through to increment each one resulting in cleaner and more efficient code.