Ai Tourist v13
The main change here was that for speech to text, I initially saved the recording of the speech. This was not not required and made the program less efficient without any benefit. I also simplified my code and removed any unnecessary functions that I only utilized once. Also played around with some mkcert certificate stuff and I got text to speech and speech to text working on my phone but only on firefox